+++ src/tabs/TabEquipment.tsx
@@ -44,13 +44,13 @@ export const TabEquipment = (): React.ReactElement => {
 				<Grid2
 					container
 					spacing={2}
-					sx={{ color: "black", justifyContent: "center" }}
+					sx={{ color: "black", justifyContent: "center", display: "flex" }}
 				>
 					<Grid2
 						key="label-equipment"
 						xs={12}
-						sm={4}
-						md={3}
+						sm={10}
+						md={10}
 						sx={{ display: "flex", flexDirection: "row" }}
 					>
 						<Item
@@ -60,7 +60,7 @@ export const TabEquipment = (): React.ReactElement => {
 								alignItems: "center",
 							}}
 						>
-							Libellé Équipement :
+							Libellé de l'équipement :
 						</Item>
 						<TextField
 							value={newLabelEquipment}
@@ -70,13 +70,17 @@ export const TabEquipment = (): React.ReactElement => {
 							}}
 							sx={{
 								m: 1,
-								flexGrow: 1,
+								flexGrow: 10,
 								backgroundColor: "white",
+								width: "800px",
 							}}
 						></TextField>
 					</Grid2>
 				</Grid2>
-				<Grid2 container sx={{ justifyContent: "center" }}>
+				<Grid2
+					container
+					sx={{ display: "flex", flexGrow: 1, justifyContent: "center" }}
+				>
 					<Button
 						id="validation-button"
 						disabled={!(newLabelEquipment !== "")}
@@ -125,10 +129,12 @@ export const TabEquipment = (): React.ReactElement => {
 	};
 
 	const handleDeleteClick = (id: GridRowId) => () => {
-		const currentRow = rows.find((row) => row.id === id);
-		const idCurrentRow = currentRow?.id;
-		axios.delete(baseURL + "/api/equipment/" + idCurrentRow);
-		setRows(rows.filter((row) => row.id !== id));
+		if (window.confirm("Voulez-vous vraiment supprimer cet équipement ?")) {
+			const currentRow = rows.find((row) => row.id === id);
+			const idCurrentRow = currentRow?.id;
+			axios.delete(baseURL + "/api/equipment/" + idCurrentRow);
+			setRows(rows.filter((row) => row.id !== id));
+		}
 	};
 
 	const handleCancelClick = (id: GridRowId) => () => {
@@ -164,6 +170,7 @@ export const TabEquipment = (): React.ReactElement => {
 			headerName: "Libellé équipement",
 			type: "string",
 			editable: true,
+			flex: 1,
 		},
 		{
 			field: "actions",
@@ -220,6 +227,7 @@ export const TabEquipment = (): React.ReactElement => {
 					checkboxSelection={false}
 					density="comfortable"
 					editMode="row"
+					getRowHeight={() => "auto"}
 					rowModesModel={rowModesModel}
 					onRowModesModelChange={(newModel) => setRowModesModel(newModel)}
 					onRowEditStart={handleRowEditStart}
